Fix Service[oddjobd] failing to start
Without this patch the initial Puppet run fails with the following error
message:

    Error: Could not start Service[oddjobd]: Execution of '/sbin/service oddjobd
    start' returned 1: Starting oddjobd: [FAILED]

The cause of this error is that the `messagebus` service is not started.  This
patch addresses the problem by adding optional management of dependent services
to the sssd::services class.  Since these dependencies aren't strictly related
to sssd they can easily be overridden by setting sssd::service_dependencies to
an empty array in Hiera.