Test only the latest minor release for major releases
We are no longer testing against each minor release in Puppet v5. You
are encouraged to keep up with the latest minor release as this is meant
to be backward compatible and safe.

By not testing each minor release, we greatly speed up our testing as
well and make the code more maintainable, especially in regards to
dealing with ruby gem dependencies.