Add support for Fedora versions 29 and 30
Add support for authselect enabled Fedora 29 and 30. Backwards
compatibility is maintained for Fedora 27 and older, which used
authconfig.

New unittests for Fedora 29 and 30 have been added. Conditional tests
previously based on the "osfamily" fact evaluating to RedHat have been
converted to use "os.name", which evaluates differently for RedHat and
Fedora operating systems.

Acceptance tests for Fedora 29 and 30 have been added. At the time of
this commit an official puppet-agent RPM for Fedora 30 does not
exist. The Fedora 30 acceptance tests install the Fedora 29
puppet-agent in a Fedora 30 Docker container. This requires the
installation of the libxcrypt compatibility package.