Remove support for end of life Fedora versions older than 30
The Fedora project keeps two versions in support. At the time of this
commit those versions are 32 and 33. However, reality is that end of
life (EOL) versions do linger. This commit removes support for Fedora
version 29 and earlier. To provided context, Fedora 29 was designated
EOL on 2019-11-26. This leaves the following two most recent EOL
versions supported by this module: 30 (EOL 2020-05-26) and 31 (EOL
2020-11-24).